add more casts

+++ mpls_bpf_kern.c
@@ -221,8 +221,8 @@ SEC("mpls_encap") int mpls_encap(struct __sk_buff *skb) {
 
   // construct our deterministic mpls header
   struct mpls_hdr mpls = mpls_encode(MPLS_STATIC_LABEL, 123, 0, true);
-  unsigned long offset = sizeof(struct ethhdr) + (unsigned long)iph_len;
-  ret = bpf_skb_store_bytes(skb, offset, &mpls, sizeof(struct mpls_hdr),
+  unsigned long offset = sizeof(struct ethhdr) + (unsigned long) iph_len;
+  ret = bpf_skb_store_bytes(skb, (int) offset, &mpls, sizeof(struct mpls_hdr),
                             BPF_F_RECOMPUTE_CSUM);
 
   return TC_ACT_OK;
